Reorg js files to be more logical + understandable
Fix bug where totals don't calculate when days per week is filled in
