+++ test/unit_tests/apps/youth-employment-success/js/models/route-spec.js
@@ -1,4 +1,4 @@
-import createRoute from '../../../../../cfgov/unprocessed/apps/youth-employment-success/js/route';
+import createRoute from '../../../../../../cfgov/unprocessed/apps/youth-employment-success/js/models/route';
 
 describe( 'route factory function', () => {
   it( 'constructs a route object', () => {
