+++ test/unit_tests/apps/youth-employment-success/js/views/budget/budget-form-spec.js
@@ -1,6 +1,6 @@
-import { simulateEvent } from '../../../../util/simulate-event';
-import budgetFormView from '../../../../../cfgov/unprocessed/apps/youth-employment-success/js/budget-form-view';
-import mockStore from '../../../mocks/store';
+import { simulateEvent } from '../../../../../../util/simulate-event';
+import budgetFormView from '../../../../../../../cfgov/unprocessed/apps/youth-employment-success/js/views/budget';
+import mockStore from '../../../../../mocks/store';
 
 const HTML = `
 <section class="block o-yes-budget">
