+++ test/unit_tests/apps/youth-employment-success/js/views/goals/goals-spec.js
@@ -1,11 +1,11 @@
-import { simulateEvent } from '../../../../../util/simulate-event';
-import goalsView from '../../../../../../cfgov/unprocessed/apps/youth-employment-success/js/views/goals';
+import { simulateEvent } from '../../../../../../util/simulate-event';
+import goalsView from '../../../../../../../cfgov/unprocessed/apps/youth-employment-success/js/views/goals';
 import {
   updateGoalAction,
   updateGoalImportanceAction,
   updateGoalStepsAction,
   updateGoalTimelineAction
-} from '../../../../../../cfgov/unprocessed/apps/youth-employment-success/js/reducers/goal-reducer';
+} from '../../../../../../../cfgov/unprocessed/apps/youth-employment-success/js/reducers/goal-reducer';
 
 const HTML = `
   <form class="js-yes-goals">
