+++ test/unit_tests/apps/youth-employment-success/js/views/route/form-questions/miles-spec.js
@@ -1,59 +1,49 @@
-import { simulateEvent } from '../../../../../util/simulate-event';
-import daysPerWeekView from '../../../../../../cfgov/unprocessed/apps/youth-employment-success/js/views/days-per-week';
+import { simulateEvent } from '../../../../../../../util/simulate-event';
+import milesView from '../../../../../../../../cfgov/unprocessed/apps/youth-employment-success/js/views/route/form-questions/miles';
+import mockStore from '../../../../../../mocks/store';
 import {
-  clearDaysPerWeekAction,
-  updateDaysPerWeekAction,
-  updateDaysToActionPlan
-} from '../../../../../../cfgov/unprocessed/apps/youth-employment-success/js/reducers/route-option-reducer';
-import TODO_FIXTURE from '../../fixtures/todo-alert';
-import TodoNotificationMock from '../../mocks/todo-notification';
+  clearMilesAction,
+  updateMilesAction,
+  updateMilesToActionPlan
+} from '../../../../../../../../cfgov/unprocessed/apps/youth-employment-success/js/reducers/route-option-reducer';
+import { PLAN_TYPES } from '../../../../../../../../cfgov/unprocessed/apps/youth-employment-success/js/data-types/todo-items';
+import TODO_FIXTURE from '../../../../fixtures/todo-alert';
+import TodoNotificationMock from '../../../../mocks/todo-notification';
 
 const HTML = `
-  <div class="m-yes-days-per-week">
-    <input type="text" name="daysPerWeek" data-js-name="daysPerWeek">
-    <input type="checkbox" name="yes-route-days-unsure">
+  <div class="m-yes-miles">
+    <input type="text">
+    <input type="checkbox">
   </div>
   ${ TODO_FIXTURE }
 `;
 
-describe( 'DaysPerWeekView', () => {
+describe( 'milesView', () => {
   const routeIndex = 0;
-  const CLASSES = daysPerWeekView.CLASSES;
-  const dispatch = jest.fn();
   const todoNotification = new TodoNotificationMock();
-  function mockStore() {
-    let subscriberFn;
-
-    return {
-      subscriber() {
-        return subscriberFn;
-      },
-      dispatch,
-      subscribe: jest.fn().mockImplementation( fn => {
-        subscriberFn = fn;
-      } )
-    };
-  }
   let store;
-  let view;
   let dom;
+  let view;
 
   beforeEach( () => {
-    document.body.innerHTML = HTML;
     store = mockStore();
-    dom = document.querySelector( `.${ CLASSES.CONTAINER }` );
-    view = daysPerWeekView( dom, { store, routeIndex, todoNotification } );
+    document.body.innerHTML = HTML;
+    dom = document.querySelector( `.${ milesView.CLASSES.CONTAINER }` );
+    view = milesView( dom, { store, routeIndex, todoNotification } );
     view.init();
   } );
 
   afterEach( () => {
     todoNotification.mockReset();
-    dispatch.mockReset();
+    store.mockReset();
     view = null;
-    store = null;
   } );
 
   describe( 'on initialize', () => {
+    it( 'hides its container element', () => {
+      expect( dom.classList.contains( 'u-hidden' ) ).toBeTruthy();
+    } );
+
     it( 'subscribes to the store', () => {
       expect( store.subscribe ).toHaveBeenCalled();
     } );
@@ -64,36 +54,35 @@ describe( 'DaysPerWeekView', () => {
   } );
 
   describe( 'event handling', () => {
-    it( 'calls the correct action on text input', () => {
-      const daysEl = dom.querySelector( 'input[type="text"]' );
-      const daysPerWeek = '2';
+    it( 'calls the correct action on `miles` field input', () => {
+      const milesEl = dom.querySelector( 'input[type="text"]' );
+      const milesPerDay = '5';
       const mock = store.dispatch.mock;
 
-      daysEl.value = daysPerWeek;
+      milesEl.value = milesPerDay;
 
-      simulateEvent( 'input', daysEl );
+      simulateEvent( 'input', milesEl );
 
       expect( mock.calls.length ).toBe( 1 );
       expect( mock.calls[0][0] ).toEqual(
-        updateDaysPerWeekAction( {
+        updateMilesAction( {
           routeIndex,
-          value: daysPerWeek
+          value: milesPerDay
         } )
       );
     } );
 
-    it( 'calls the correct action on checkbox input', () => {
-      const notSureEl = dom.querySelector( 'input[type="checkbox"]' );
+    it( 'calls the correct action when the `not sure` checkbox is clicked', () => {
+      const notSureEl = document.querySelector( 'input[type="checkbox"]' );
       const mock = store.dispatch.mock;
 
-      notSureEl.checked = 'false';
       simulateEvent( 'click', notSureEl );
 
       expect( mock.calls.length ).toBe( 1 );
       expect( mock.calls[0][0] ).toEqual(
-        updateDaysToActionPlan( {
+        updateMilesToActionPlan( {
           routeIndex,
-          value: false
+          value: true
         } )
       );
     } );
@@ -116,12 +105,11 @@ describe( 'DaysPerWeekView', () => {
   } );
 
   describe( 'on state update', () => {
-    it( 'removes u-hidden class when transportation mode is drive', () => {
+    it( 'removes u-hidden class when transportation mode is Drive', () => {
       const state = {
         routes: {
           routes: [ {
-            transportation: 'Drive',
-            actionPlanItems: []
+            transportation: 'Drive'
           } ]
         }
       };
@@ -131,45 +119,41 @@ describe( 'DaysPerWeekView', () => {
       expect( dom.classList.contains( 'u-hidden' ) ).toBeFalsy();
     } );
 
-    describe( 'when average cost is defined as a monthly cost', () => {
-      const days = '2';
-      const checked = 'checked';
+    describe( 'when transportation mode is not Drive', () => {
+      const miles = '12';
+      const checked = true;
       const prevState = {
         routes: {
           routes: [ {
-            daysPerWeek: days,
-            actionPlanItems: []
+            miles: miles,
+            actionPlanItems: [ PLAN_TYPES.MILES ]
           } ]
         }
       };
       const state = {
         routes: {
           routes: [ {
-            transportation: 'Walk',
-            isMonthlyCost: true,
-            actionPlanItems: []
+            transportation: 'Walk'
           } ]
         }
       };
 
       it( 'clears the form inputs', () => {
-        const daysEl = dom.querySelector( 'input[type="text"]' );
+        const milesEl = dom.querySelector( 'input[type="text"]' );
         const notSureEl = dom.querySelector( 'input[type="checkbox"]' );
 
-        daysEl.value = days;
+        milesEl.value = miles;
         notSureEl.checked = checked;
 
         store.subscriber()( prevState, state );
 
-        expect( daysEl.value ).toBe( '' );
+        expect( milesEl.value ).toBe( '' );
         expect( notSureEl.checked ).toBe( false );
       } );
 
-      it( 'hides the container element, if not hidden', () => {
+      it( 'hides the container element', () => {
         const subscriberFn = store.subscriber();
         subscriberFn( prevState, state );
-
-        expect( dom.classList.contains( 'u-hidden' ) ).toBeTruthy();
 
         subscriberFn( { routes: { routes: [ {} ]}}, {
           routes: {
@@ -186,18 +170,34 @@ describe( 'DaysPerWeekView', () => {
         expect( dom.classList.contains( 'u-hidden' ) ).toBeTruthy();
       } );
 
-      it( 'dispatches the correct action when daysPerWeek is filled in', () => {
+      it( 'dispatches the correct action when miles is filled in and transportation method is not drive', () => {
         const mock = store.dispatch.mock;
         store.subscriber()( prevState, state );
 
         expect( mock.calls.length ).toBe( 1 );
         expect( mock.calls[0][0] ).toEqual(
-          clearDaysPerWeekAction( { routeIndex } )
+          clearMilesAction( { routeIndex } )
+        );
+      } );
+
+      it( 'dispatches the correct action when not sure is selected and transportation method is not drive', () => {
+        const mock = store.dispatch.mock;
+        store.subscriber()( {
+          routes: {
+            routes: [ {
+              actionPlanItems: [ PLAN_TYPES.MILES ]
+            } ]
+          }
+        }, state );
+
+        expect( mock.calls.length ).toBe( 1 );
+        expect( mock.calls[0][0] ).toEqual(
+          clearMilesAction( { routeIndex } )
         );
       } );
 
       it( 'calls .remove on the todo notification component when this view is toggled', () => {
-        store.subscriber()( { routes: { routes: [ {} ]}}, state );
+        store.subscriber()( prevState, state );
 
         expect( todoNotification.remove.mock.calls.length ).toBe( 1 );
       } );
