+++ test/unit_tests/apps/youth-employment-success/js/views/todo-notification-spec.js
@@ -1,5 +1,5 @@
-import TodoNotification from '../../../../../cfgov/unprocessed/apps/youth-employment-success/js/todo-notification';
-import TODO_FIXTURE from '../fixtures/todo-alert';
+import TodoNotification from '../../../../../../cfgov/unprocessed/apps/youth-employment-success/js/views/todo-notification';
+import TODO_FIXTURE from '../../fixtures/todo-alert';
 
 const CONTAINER = 'alert-container';
 const HTML = `
